Extract test suite finalized integration tester
